Attach template names to issuers
The charts were updated to accept a pcaTemplateName that maps to a
template ARN. There is a new example showing this. The template mapping
was updated to reflect this. The controllers were adjusted to call the
updated function. The unit tests were updated to validate the
BuildTemplateArn function,

Change templateArn to pcaTemplateName in crds